feat: add test case for refinery price list and fg bom to the workflow
