+++ .github/workflows/core-tests.yml
@@ -55,3 +55,5 @@ jobs:
           bench --site test_site run-tests --app jewellery_erpnext --doctype "Employee IR"
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_employee_ir_loss_baseline
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_warehouse_stock_entry
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_fg_bom_field_configuration
+          bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_repair_and_fg_bom
