+++ .github/workflows/refining-tests.yml
@@ -48,6 +48,7 @@ jobs:
           cd ~/frappe-bench
           echo "Running Refining & Advanced Tests..."
           bench --site test_site run-tests --app jewellery_erpnext --doctype "Refining Entry"
+          bench --site test_site run-tests --app jewellery_erpnext --doctype "Refinery Price List"
           bench --site test_site run-tests --app jewellery_erpnext --doctype "Product Certification"
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_serialize
           bench --site test_site run-tests --module jewellery_erpnext.jewellery_erpnext.tests.test_lock_order
